+++ lib/ripple_container.dart
@@ -5,9 +5,6 @@ library ripple_container;
 
 import 'package:flutter/material.dart';
 
-import 'container_actions.dart';
-import 'effect_options.dart';
-
 class RippleContainer extends StatelessWidget {
   /// [child] is required parameter.
   ///specify the widget to be displayed within the RippleContainer.
@@ -23,14 +20,54 @@ class RippleContainer extends StatelessWidget {
   /// Sets the background color of the RippleContainer.
   /// If null is entered, transparent color is the default.
   final Color backgroundColor;
+  final BlendMode? backgroundBlendMode;
 
-  /// [decoration] is RippleContainer's decoration
-  final Decoration? decoration;
+  final Gradient? gradient;
 
-  final EffectOptions? effectOptions;
+  /// The [borderRadius] determines the border radius of the Container,
+  /// and the child is clipped to fit the Container.
+  /// default value is [BorderRadius.zero]
+  final BorderRadius borderRadius;
+  final BoxBorder? border;
 
-  final ContainerActions? containerActions;
+  /// [padding] is applied inside the Container,
+  /// while [margin] is the value used to create space outside the Container.
+  /// default value is [EdgeInsets.zero].
+  final EdgeInsets padding;
+  final EdgeInsets margin;
 
+  /// The [boxShadow] variable is used to apply a shadow outside the Container.
+  final List<BoxShadow>? boxShadow;
+
+  /// Defines the appearance of the splash.
+  ///
+  /// Defaults to the value of the theme's splash factory: [ThemeData.splashFactory].
+  ///
+  /// See also:
+  ///
+  ///  * [radius], the (maximum) size of the ink splash.
+  ///  * [splashColor], the color of the splash.
+  ///  * [highlightColor], the color of the highlight.
+  ///  * [InkSplash.splashFactory], which defines the default splash.
+  ///  * [InkRipple.splashFactory], which defines a splash that spreads out
+  ///    more aggressively than the default.
+  final InteractiveInkFeatureFactory? splashFactory;
+
+  /// The splash color of the ink response. If this property is null then the
+  /// splash color of the theme, [ThemeData.splashColor], will be used.
+  ///
+  /// See also:
+  ///
+  ///  * [splashFactory], which defines the appearance of the splash.
+  ///  * [radius], the (maximum) size of the ink splash.
+  ///  * [highlightColor], the color of the highlight.
+  final Color? splashColor;
+
+
+  /// The [onTap] and [onLongPress] variables are used to define the actions when this widget is selected.
+  /// If both variables are null, there will be no splash effect.
+  final Function()? onTap;
+  final Function()? onLongPress;
 
   const RippleContainer({
     super.key,
@@ -38,75 +75,51 @@ class RippleContainer extends StatelessWidget {
     this.width,
     this.height,
     this.backgroundColor = Colors.transparent,
-    this.decoration,
-    this.effectOptions,
-    this.containerActions,
+    this.borderRadius = BorderRadius.zero,
+    this.padding = EdgeInsets.zero,
+    this.margin = EdgeInsets.zero,
+    this.backgroundBlendMode,
+    this.boxShadow,
+    this.border,
+    this.gradient,
+    this.splashFactory,
+    this.splashColor,
+    this.onTap,
+    this.onLongPress,
   });
 
   @override
   Widget build(BuildContext context) {
-    /// Method for preventing overlapping use conflict between color and backgroundColor in BoxDecoration.
-    final reinterpretationDeco = reinterpretationDecoration(decoration, backgroundColor);
-
     return Container(
-      width: width,
-      height: height,
-      decoration: reinterpretationDeco,
-      color: (reinterpretationDeco is BoxDecoration) ? null : backgroundColor,
-      alignment: Alignment.center,
-      child: Material(
-        color: Colors.transparent,
-        child: InkWell(
-          ///Actions
-          onTap: containerActions?.onTap ?? () {},
-          onTapCancel: containerActions?.onTapCancel,
-          onTapDown: containerActions?.onTapDown,
-          onTapUp: containerActions?.onTapUp,
-          onSecondaryTap: containerActions?.onSecondaryTap,
-          onSecondaryTapCancel: containerActions?.onSecondaryTapCancel,
-          onSecondaryTapDown: containerActions?.onSecondaryTapDown,
-          onSecondaryTapUp: containerActions?.onSecondaryTapUp,
-          onDoubleTap: containerActions?.onDoubleTap,
-          onLongPress: containerActions?.onLongPress,
-          mouseCursor: containerActions?.mouseCursor,
-          onHighlightChanged: containerActions?.onHighlightChanged,
-          onHover: containerActions?.onHover,
-          ///InkGraphic Options
-          borderRadius: effectOptions?.borderRadius,
-          autofocus: effectOptions?.autofocus ?? false,
-          canRequestFocus: effectOptions?.canRequestFocus ?? true,
-          customBorder: effectOptions?.customBorder,
-          enableFeedback: effectOptions?.enableFeedback,
-          excludeFromSemantics: effectOptions?.excludeFromSemantics ?? false,
-          focusColor: effectOptions?.focusColor,
-          focusNode: effectOptions?.focusNode,
-          highlightColor: effectOptions?.highlightColor,
-          hoverColor: effectOptions?.hoverColor,
-          hoverDuration: effectOptions?.hoverDuration,
-          onFocusChange: effectOptions?.onFocusChange,
-          overlayColor: effectOptions?.overlayColor,
-          radius: effectOptions?.radius,
-          splashColor: effectOptions?.splashColor,
-          splashFactory: effectOptions?.splashFactory,
-          statesController: effectOptions?.statesController,
-          child: Container(
-            width: width,
-            height: height,
-            alignment: Alignment.center,
-            child: child,
+      margin: margin,
+      decoration: BoxDecoration(
+        color: backgroundColor,
+        borderRadius: borderRadius,
+        boxShadow: boxShadow,
+        backgroundBlendMode: backgroundBlendMode,
+        border: border,
+        gradient: gradient,
+      ),
+      child: ClipRRect(
+        borderRadius: borderRadius,
+        child: Material(
+          color: Colors.transparent,
+          child: InkWell(
+            splashFactory: splashFactory,
+            splashColor: splashColor,
+            borderRadius: borderRadius,
+            onTap: onTap,
+            onLongPress: onLongPress,
+            child: Container(
+              width: width,
+              height: height,
+              padding: padding,
+              alignment: Alignment.center,
+              child: child,
+            ),
           ),
         ),
       ),
     );
-  }
-
-  // Method for preventing overlapping use conflict between color and backgroundColor in BoxDecoration.
-  Decoration? reinterpretationDecoration(Decoration? decoration, Color backgroundColor) {
-    Decoration? resultDecoration = decoration;
-    if (decoration is BoxDecoration) {
-      final BoxDecoration boxDecoration = decoration;
-      resultDecoration = boxDecoration.copyWith(color: backgroundColor);
-    }
-    return resultDecoration;
   }
 }