+++ CHANGELOG.md
@@ -1,3 +1,27 @@
 ## 0.0.1
 
 * Test Version
+
+## 0.1.0
+
+* overridden the interface.
+```
+const RippleContainer({
+    super.key,
+    required this.child,
+    this.width,
+    this.height,
+    this.backgroundColor = Colors.transparent,
+    this.borderRadius = BorderRadius.zero,
+    this.padding = EdgeInsets.zero,
+    this.margin = EdgeInsets.zero,
+    this.backgroundBlendMode,
+    this.boxShadow,
+    this.border,
+    this.gradient,
+    this.splashFactory,
+    this.splashColor,
+    this.onTap,
+    this.onLongPress,
+  });
+```